bottleneck ir module reading and writing
