+++ source/slang/slang-serialize-ir.h
@@ -13,108 +13,22 @@
 
 namespace Slang
 {
-struct IRModuleChunk : RIFF::ListChunk
-{
-};
 
-struct IRSerialWriter
-{
-    typedef IRSerialData Ser;
-    typedef IRSerialBinary Bin;
+struct IRModule;
 
-    Result write(
-        IRModule* module,
-        SerialSourceLocWriter* sourceLocWriter,
-        IRSerialData* serialData);
+void writeSerializedModuleIR(
+    RIFF::BuildCursor& cursor,
+    IRModule* moduleDecl,
+    SerialSourceLocWriter* sourceLocWriter);
 
-    /// Write to a container
-    static Result writeTo(const IRSerialData& data, RIFF::BuildCursor& cursor);
-
-    /// Get an instruction index from an instruction
-    Ser::InstIndex getInstIndex(IRInst* inst) const
-    {
-        return inst ? Ser::InstIndex(m_instMap.getValue(inst)) : Ser::InstIndex(0);
-    }
-
-    /// Get a slice from an index
-    UnownedStringSlice getStringSlice(Ser::StringIndex index) const
-    {
-        return m_stringSlicePool.getSlice(StringSlicePool::Handle(index));
-    }
-    /// Get index from string representations
-    Ser::StringIndex getStringIndex(StringRepresentation* string)
-    {
-        return Ser::StringIndex(m_stringSlicePool.add(string));
-    }
-    Ser::StringIndex getStringIndex(const UnownedStringSlice& slice)
-    {
-        return Ser::StringIndex(m_stringSlicePool.add(slice));
-    }
-    Ser::StringIndex getStringIndex(Name* name)
-    {
-        return name ? getStringIndex(name->text) : SerialStringData::kNullStringIndex;
-    }
-    Ser::StringIndex getStringIndex(const char* chars)
-    {
-        return Ser::StringIndex(m_stringSlicePool.add(chars));
-    }
-    Ser::StringIndex getStringIndex(const String& string)
-    {
-        return Ser::StringIndex(m_stringSlicePool.add(string.getUnownedSlice()));
-    }
-
-    StringSlicePool& getStringPool() { return m_stringSlicePool; }
-
-    IRSerialWriter()
-        : m_serialData(nullptr), m_stringSlicePool(StringSlicePool::Style::Default)
-    {
-    }
-
-    /// Produces an instruction list which is in same order as written through IRSerialWriter
-    static void calcInstructionList(IRModule* module, List<IRInst*>& instsOut);
-
-protected:
-    void _addInstruction(IRInst* inst);
-    Result _calcDebugInfo(SerialSourceLocWriter* sourceLocWriter);
-
-    List<IRInst*> m_insts; ///< Instructions in same order as stored in the
-
-    List<IRDecoration*>
-        m_decorations; ///< Holds all decorations in order of the instructions as found
-    List<IRInst*> m_instWithFirstDecoration; ///< All decorations are held in this order after all
-                                             ///< the regular instructions
-
-    Dictionary<IRInst*, Ser::InstIndex> m_instMap; ///< Map an instruction to an instruction index
-
-    StringSlicePool m_stringSlicePool;
-    IRSerialData* m_serialData; ///< Where the data is stored
-};
-
-struct IRSerialReader
-{
-    typedef IRSerialData Ser;
-
-    /// Read a stream to fill in dataOut IRSerialData
-    static Result readFrom(IRModuleChunk const* irModuleChunk, IRSerialData* outData);
-
-    /// Read a module from serial data
-    Result read(
-        const IRSerialData& data,
-        Session* session,
-        SerialSourceLocReader* sourceLocReader,
-        RefPtr<IRModule>& outModule);
-
-    IRSerialReader()
-        : m_serialData(nullptr), m_module(nullptr), m_stringTable(StringSlicePool::Style::Default)
-    {
-    }
-
-protected:
-    StringSlicePool m_stringTable;
-
-    const IRSerialData* m_serialData;
-    IRModule* m_module;
-};
+[[nodiscard]]
+SlangResult readSerializedModuleIR(
+    RIFF::Chunk const* chunk,
+    // ISlangBlob* blobHoldingSerializedData,
+    Session* session,
+    // DiagnosticSink* sink,
+    SerialSourceLocReader* sourceLocReader,
+    RefPtr<IRModule>& outIRModule);
 
 } // namespace Slang
